implemented authentication

+++ router/endpoints.js
@@ -4,14 +4,15 @@ const router = express.Router();
 
 const mongodb = require('mongodb');
 const { ObjectID } = mongodb;
-
 const { getCollection } = require('../util/db');
 const endpointsCollection = getCollection('endpoints');
 
-router.post('/add', async (req, res) => {
+const { tokenMiddlware } = require('../middleware/auth');
+
+router.post('/add', tokenMiddlware, async (req, res) => {
   try {
     const body = req.body;
-    const result = await endpointsCollection.insertOne(body);
+    const result = await endpointsCollection.insertOne({ ...body, user_id: req.user._id });
     const { insertedId } = result;
     res.status(201).json({ message: "endpoint inserted successfuly", insertedId });
   } catch (err) {
@@ -20,9 +21,9 @@ router.post('/add', async (req, res) => {
   }
 });
 
-router.get('/get-all', async (req, res) => {
+router.get('/get-all', tokenMiddlware, async (req, res) => {
   try {
-    const result = await endpointsCollection.find().toArray();
+    const result = await endpointsCollection.find({ _id: req.user._id }).toArray();
     res.send(result);
   } catch (err) {
     console.log(err);
@@ -30,10 +31,10 @@ router.get('/get-all', async (req, res) => {
   }
 });
 
-router.get('/get/:id', async (req, res) => {
+router.get('/get/:id', tokenMiddlware, async (req, res) => {
   try {
     const { id } = req.params;
-    const result = await endpointsCollection.findOne({ _id: new ObjectID((`${id}`)) });
+    const result = await endpointsCollection.findOne({ $and: [{ user_id: req.user._id }, { _id: new ObjectID((`${id}`)) }] });
     res.send(result);
   } catch (err) {
     console.log(err);
@@ -41,10 +42,10 @@ router.get('/get/:id', async (req, res) => {
   }
 });
 
-router.post('/get', async (req, res) => {
+router.post('/get', tokenMiddlware, async (req, res) => {
   try {
     const { body } = req;
-    const result = await endpointsCollection.find(body).toArray();
+    const result = await endpointsCollection.find({ $and: [{ user_id: req.user._id }, { ...body }] }).toArray();
     res.send(result);
   } catch (error) {
     console.log(err);
@@ -52,12 +53,12 @@ router.post('/get', async (req, res) => {
   }
 });
 
-router.patch('/update', async (req, res) => {
+router.patch('/update', tokenMiddlware, async (req, res) => {
   try {
     const { body } = req;
     const id = new ObjectID(body._id);
     delete body._id;
-    await endpointsCollection.updateOne({ _id: id }, { $set: body });
+    await endpointsCollection.updateOne({ $and: [{ user_id: req.user._id }, { _id: id }] }, { $set: body });
     res.json({ message: "updated successfully" });
   } catch (err) {
     console.log(err);
@@ -65,13 +66,13 @@ router.patch('/update', async (req, res) => {
   }
 });
 
-router.put('/replace', async (req, res) => {
+router.put('/replace', tokenMiddlware, async (req, res) => {
   try {
     const { body } = req;
     if (!body._id) return res.status(403).json({ message: "_id is not provided!" });
     const id = new ObjectID(body._id);
     body._id = id;
-    await endpointsCollection.replaceOne({ _id: id }, body);
+    await endpointsCollection.replaceOne({ $and: [{ user_id: req.user._id }, { _id: id }] }, { ...body, user_id: req.user._id });
     res.json({ message: "replaced successfully" });
   } catch (err) {
     console.log(err);
@@ -79,11 +80,11 @@ router.put('/replace', async (req, res) => {
   }
 });
 
-router.delete('/delete/:id', async (req, res) => {
+router.delete('/delete/:id', tokenMiddlware, async (req, res) => {
   try {
     const { id } = req.params;
     const _id = new ObjectID(`${id}`);
-    await endpointsCollection.deleteOne({ _id });
+    await endpointsCollection.deleteOne({ $and: [{ user_id: req.user._id }, { _id }] });
     res.json({ message: "deleted successfuly" });
   } catch (err) {
     console.log(err);
